pyproject.toml
[project]
name = "sl0thm4n-image-preprocessor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "numpy>=2.3.4",
    "onnxruntime-gpu>=1.23.0",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "rembg>=2.0.67",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-black>=0.6.0",
    "pytest-cov>=7.0.0",
    "pytest-flake8>=1.3.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--black --flake8 --cov=sl0thfier --cov-report=term-missing -v"
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
select = ["E", "F"]
exclude = ["build", "dist", ".venv", "__pycache__"]
extend-ignore = ["E203"]
